Porting Update 8
I changed the compiler used when running tests becasue the VS compiler supports the syntax from newer versions of C#. I also added the required dependency for compilation and got rid of the one for RestSharp. In addition, file data is now inserted into the request body if a file is being sent with the request.